requirements.txt
flask
requests
pyyaml
watchdog
